+++ AutomatedLabUnattended/Private/CloudInit/Add-UnattendedCloudInitRenameNetworkAdapters.ps1
@@ -1,0 +1,9 @@
+﻿function Add-UnattendedCloudInitRenameNetworkAdapters
+{
+	[CmdletBinding()]
+    param
+    (
+    )
+	
+    Write-PSFMessage -Message 'Method not required on Ubuntu/Cloudinit'
+}