+++ AutomatedLabUnattended/Private/CloudInit/Add-UnattendedCloudInitSynchronousCommand.ps1
@@ -1,0 +1,13 @@
+﻿function Add-UnattendedCloudInitSynchronousCommand
+{
+	[CmdletBinding()]
+    param (
+        [Parameter(Mandatory)]
+        [string]$Command,
+
+        [Parameter(Mandatory)]
+        [string]$Description
+    )
+
+    $script:un['late-commands'] += $Command
+}