+++ AutomatedLabUnattended/Private/CloudInit/Set-UnattendedCloudInitAdministratorName.ps1
@@ -1,0 +1,11 @@
+﻿function Set-UnattendedCloudInitAdministratorName
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [string]
+        $Name
+    )
+
+    $Script:un.identity.username = $Name
+}