+++ AutomatedLabUnattended/Private/CloudInit/Set-UnattendedCloudInitComputerName.ps1
@@ -1,0 +1,11 @@
+﻿function Set-UnattendedCloudInitComputerName
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [string]
+        $ComputerName
+    )
+
+    $Script:un.identity.hostname = $ComputerName
+}