+++ AutomatedLabUnattended/Private/CloudInit/Set-UnattendedCloudInitProductKey.ps1
@@ -1,0 +1,10 @@
+﻿function Set-UnattendedProductKey
+{
+	[CmdletBinding()]
+	param (
+		[Parameter(Mandatory = $true)]
+		[string]$ProductKey
+	)
+
+	Write-PSFMessage "No product key required on CloudInit/Ubuntu"
+}