+++ AutomatedLabUnattended/Private/RedHat/Set-UnattendedKickstartDomain.ps1
@@ -14,5 +14,13 @@
 		[string]$OrganizationalUnit
     )
 
-	$script:un.Add(("realm join --one-time-password='{0}' {1}" -f $Password, $DomainName))
+	if ($OrganizationalUnit)
+	{
+		$script:un.Add(("realm join --computer-ou='{2}' --one-time-password='{0}' {1}" -f $Password, $DomainName, $OrganizationalUnit))
+
+	}
+	else
+	{
+		$script:un.Add(("realm join --one-time-password='{0}' {1}" -f $Password, $DomainName))
+	}
 }